refactor(gateway-core): sink namespace pure helpers (#1368)
Move the bare-name resolver, tool-name codec, and naming constants from
`dcc-mcp-gateway::gateway::namespace::{bare,constants,encode}` into the
inner `dcc-mcp-gateway-core::naming::*` module. The original paths stay
as thin facades that re-export from the core crate, so all historical
`dcc_mcp_gateway::namespace::*` imports keep compiling.

Splits the moved code along SRP boundaries:

* `naming/primitives.rs` — low-level encoding constants and UUID helpers
* `naming/constants.rs` — tool-name lists and separators
* `naming/encode.rs` — skill and cursor-safe tool name codec
* `naming/bare.rs` — pure bare-name collision resolver
* `naming/observability.rs` — one-shot warn state for collisions
* `naming/mod.rs` — public API orchestrator

Drops the now-unused `dcc-mcp-gateway` dependency from
`dcc-mcp-http-server`; the runtime layer reaches only into
`dcc-mcp-gateway-core` for these helpers, which removes the prior
reverse dependency from the runtime tier onto the aggregation tier and
unblocks the Phase 1 crate split tracked in #1357.

Refs #1367.
