+++ crates/dcc-mcp-gateway-core/src/naming/tests.rs
