+++ crates/dcc-mcp-http-server/src/mcp_tool_list_builder.rs
@@ -2,7 +2,7 @@
 
 use std::collections::{BTreeMap, HashSet};
 
-use dcc_mcp_gateway::namespace::{BareNameInput, resolve_bare_names};
+use dcc_mcp_gateway_core::naming::{BareNameInput, resolve_bare_names};
 use dcc_mcp_jsonrpc::{McpTool, TOOLS_LIST_PAGE_SIZE, decode_cursor, encode_cursor};
 use dcc_mcp_naming::validate_tool_name;
